Blimp module ready for quickstart (docker-compose integration).
This commit is the first step for the development of the toolbox required in task 7.6.
The MONICA_TODO file was also updated according with the future action points.

The blimp module has a docker-compose able to instantiate the GOST server, the SCRAL Blimp module and an MQTT listener for default topic ("GOST/").

The main additions in this commits are:
- the SCRAL module has now a client id for its MQTT publisher (for the moment the catalog_name was used as identifier together with a random number);
- new docker-compose.yml file inside blimp folder;
- MONICA_TODO.md was updated;
- constants of the client_mqtt_test.py was changed according to this new configuration, the client will be improved with an appropriate settings file.

L.M.