A new method called "entity_override" was added.
This method allow to override an OGC entity, it has to be deeply tested.
Together with this method another method was developed: "entity_registration". This function is the one that execute the POST request against GOST. It is used both by entity_discovery and by entity_registration if the looking entity does not exist.