"bug-fixing": Inserted a small delay to avoid a startup issue.
Sometimes, during the OGC configuration startup, could happens that the registration of the OGC Thing (the interested platform) fails due to the fact that the OGC Location associated is not yet registered inside GOST server. To try to avoid this issue we added a small delay (sleeping 2 seconds) between the registration of the LOCATION and the registration of the THING.
Extra: the MQTT listener in docker-compose.yml was commented by default.